// src/components/library/appLogo/types.ts
export interface AppLogoInterfaceI {
    className?: string;
    height?: string;
    width?: string;
    mHeight?: string;
    mWidth?: string;
    size?: string;
    mSize?: string;
    direction?: string;
    margin?: string;
    mMargin?: string;
    hideIcon?: boolean;
    hideName?: boolean;
    styles?: string;
}

export interface TagLinePropsI {
    className?: string;
    size?: string;
    mSize?: string;
    margin?: string;
    mMargin?: string;
    color?: string;
}